Make common code reusable
